Refactor, still need coverage.
