+++ src/Makefile.in
@@ -52,6 +52,7 @@ SRCS_COMMON = \
 SRCS = \
 	command/annotate/annotate.c \
 	command/archive/common.c \
+	command/archive/find.c \
 	command/archive/get/file.c \
 	command/archive/get/get.c \
 	command/archive/get/protocol.c \
