+++ src/command/archive/common.h
@@ -94,16 +94,6 @@ FN_EXTERN bool walIsSegment(const String *walSegment);
 // Generates the location of the wal directory using a relative wal path and the supplied pg path
 FN_EXTERN String *walPath(const String *walFile, const String *pgPath, const String *command);
 
-// Find a WAL segment in the repository. The file name can have several things appended such as a hash, compression extension, and
-// partial extension so it is possible to have multiple files that match the segment, though more than one match is not a good
-// thing.
-typedef struct WalSegmentFind WalSegmentFind;
-
-FN_EXTERN WalSegmentFind *walSegmentFindNew(const Storage *storage, const String *archiveId, unsigned int total, TimeMSec timeout);
-FN_EXTERN String *walSegmentFind(WalSegmentFind *this, const String *walSegment);
-
-FN_EXTERN String *walSegmentFindOne(const Storage *storage, const String *archiveId, const String *walSegment, TimeMSec timeout);
-
 // Get the next WAL segment given a WAL segment and WAL segment size
 FN_EXTERN String *walSegmentNext(const String *walSegment, size_t walSegmentSize, unsigned int pgVersion);
