+++ src/command/archive/get/file.c
@@ -3,7 +3,7 @@ Archive Get File
 ***********************************************************************************************************************************/
 #include "build.auto.h"
 
-#include "command/archive/common.h"
+#include "command/archive/find.h"
 #include "command/archive/get/file.h"
 #include "command/control/common.h"
 #include "common/compress/helper.h"
