+++ test/define.yaml
@@ -792,6 +792,7 @@ unit:
 
         coverage:
           - command/archive/common
+          - command/archive/find
 
       # ----------------------------------------------------------------------------------------------------------------------------
       - name: archive-get
